inetmanet@fff5dfd ieee80211: cosmetics and fixes for warnings

+++ src/linklayer/ieee80211/mac/Ieee80211Mac.cc
@@ -200,7 +200,7 @@ void Ieee80211Mac::registerInterface()
     e->setInterfaceToken(address.formInterfaceIdentifier());
 
     // FIXME: MTU on 802.11 = ?
-    e->setMtu(par("mtu"));
+    e->setMtu(par("mtu").longValue());
 
     // capabilities
     e->setBroadcast(true);
